Separated out parts of SquiggleChart, and renamed files around
