+++ packages/components/src/components/SquiggleEditor.tsx
