+++ packages/components/src/components/SquigglePlayground.tsx
