+++ packages/components/src/stories/NumberShower.stories.mdx
@@ -1,4 +1,4 @@
-import { NumberShower } from "../NumberShower";
+import { NumberShower } from "../components/NumberShower";
 import { Canvas, Meta, Story, Props } from "@storybook/addon-docs";
 
 <Meta title="Squiggle/NumberShower" component={NumberShower} />
