+++ packages/components/src/stories/SquiggleChart.stories.mdx
@@ -1,4 +1,4 @@
-import { SquiggleChart } from "../SquiggleChart";
+import { SquiggleChart } from "../components/SquiggleChart";
 import { Canvas, Meta, Story, Props } from "@storybook/addon-docs";
 
 <Meta title="Squiggle/SquiggleChart" component={SquiggleChart} />
