+++ packages/components/src/stories/SquiggleEditor.stories.mdx
@@ -1,4 +1,4 @@
-import { SquiggleEditor } from "../SquiggleEditor";
+import { SquiggleEditor } from "../components/SquiggleEditor";
 import { Canvas, Meta, Story, Props } from "@storybook/addon-docs";
 
 <Meta title="Squiggle/SquiggleEditor" component={SquiggleEditor} />
