+++ packages/components/src/stories/SquigglePlayground.stories.mdx
@@ -1,4 +1,4 @@
-import SquigglePlayground from "../SquigglePlayground";
+import SquigglePlayground from "../components/SquigglePlayground";
 import { Canvas, Meta, Story, Props } from "@storybook/addon-docs";
 
 <Meta title="Squiggle/SquigglePlayground" component={SquigglePlayground} />
