+++ packages/components/tsconfig.json
@@ -16,7 +16,7 @@
     "declaration": true,
     "sourceMap": true
   },
-  "files": ["src/spec-distributions.json", "src/spec-percentiles.json"],
+  "files": ["src/vega-specs/spec-distributions.json", "src/vega-specs/spec-percentiles.json"],
   "target": "ES6",
   "include": ["src/**/*", "src/*"],
   "exclude": ["node_modules", "**/*.spec.ts", "webpack.config.js"],
